Move queries to where they belong.
